--- producer/base/error.go
@@ -1,10 +1,0 @@
-package log_producer
-
-type BufferBusy struct{}
-
-func (b BufferBusy) Error() string {
-
-	msg := "producer buffer busy, please increase producer queue buffer"
-
-	return msg
-}
