--- producer/base/package_meta.go
@@ -1,20 +1,0 @@
-package log_producer
-
-import (
-	"sync"
-	"time"
-)
-
-type PackageMeta struct {
-	Name           string
-	ArriveTimeInMS int64
-	LogLinesCount  int
-	PackageBytes   int
-	Lock           *sync.Mutex
-}
-
-func (p *PackageMeta) Clear() {
-	p.ArriveTimeInMS = time.Now().UnixNano() / (1000 * 1000)
-	p.LogLinesCount = 0
-	p.PackageBytes = 0
-}
